Update the start-queue.sh script to match what we're using on the bots.
Over time we've added more magic git commands to clean up the local repo after
it has been borked in various ways.  The idea behind this sequence is to clear
out any changes in the working copy and then to rebuild the master branch from
origin/master.

This patch also removes the watchdog timer that we've stopped using on the
bots.