REGRESSION(r138729): crash and assertion failures in WebKit::WebProcess::plugInDidReceiveUserInteraction https://bugs.webkit.org/show_bug.cgi?id=106042 <rdar://problem/12953469>
Reviewed by Brady Eidson.

Source/WebKit2:

The assertion is hit if plugin snapshotting is disabled, and the user interacts with a plug-in that
comes from an origin not found in the cached dictionary. The proper fix is to check for whether the
setting is enabled. Otherwise, that origin should be in the dictionary since the plugin is running.

* WebProcess/Plugins/PluginView.cpp:
(WebKit::PluginView::pluginDidReceiveUserInteraction): Check for the setting, and avoid the whole code
path if snapshotting is disabled.

LayoutTests:

* platform/mac-wk2/TestExpectations: Unskip the crashing tests.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@138764 268f45cc-cd09-0410-ab3c-d52691b4dbfc