REGRESSION: [Mac] Intermittent crash in WTR::AccessibilityUIElement::isEqual https://bugs.webkit.org/show_bug.cgi?id=106073
Reviewed by Ryosuke Niwa.

Check that the element being compared to is not nil. 
This crash is happening because one of the accessibility tests does an isEqual comparison within a callback function.
It looks like occasionally, one of the elements passed in to the isEqual is nil because the notification handler is on the accessibility
controller, and the notification is being delivered after the document objects are cleared. This shouldn't happen but I've seen it 
sometimes before on parallel WebKitTestRunner runs.

* DumpRenderTree/AccessibilityUIElement.cpp:
(AccessibilityUIElement::isEqual):
* WebKitTestRunner/InjectedBundle/mac/AccessibilityUIElementMac.mm:
(WTR::AccessibilityUIElement::isEqual):



git-svn-id: http://svn.webkit.org/repository/webkit/trunk@138781 268f45cc-cd09-0410-ab3c-d52691b4dbfc