PseudoElement should never dispatch events https://bugs.webkit.org/show_bug.cgi?id=106048
Reviewed by Ojan Vafai.

Source/WebCore:

The new DOM based generated content can sometimes dispatch events
as a result of animations or transitions which is incorrect as it
exposes the PseudoElement to script. All event dispatch from a
PseudoElement should be ignored.

Per discussions with Tab Atkins in the future we might add support
by making events dispatched from a pseudo element fire with the
target as the parent element and add a new pseudo property to
events that has the value "before" or "after".

Test: fast/css-generated-content/pseudo-element-events.html

* dom/PseudoElement.h:

LayoutTests:

Add a test that events don't get dispatched from pseudo elements. The
only way right now to cause this is with transitions or animations.

* fast/css-generated-content/pseudo-element-events-expected.txt: Added.
* fast/css-generated-content/pseudo-element-events.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@138832 268f45cc-cd09-0410-ab3c-d52691b4dbfc