REGRESSION(r136967): margin-top + overflow:hidden causes incorrect layout for internal floated elements https://bugs.webkit.org/show_bug.cgi?id=106374
Reviewed by David Hyatt.

Source/WebCore:

r136397 treated any self-collapsing block that had a clearance delta as though it was clearing a float, but
blocks that avoid floats can get a clearance delta too. So just ensure there is clearance on the block when deciding
whether we need to add the margin back in before placing the float.

Test: fast/block/margin-collapse/self-collapsing-block-with-overflow-hidden-and-float-child.html

* rendering/RenderBlockLineLayout.cpp:
(WebCore::RenderBlock::LineBreaker::skipLeadingWhitespace):

LayoutTests:

* fast/block/margin-collapse/self-collapsing-block-with-overflow-hidden-and-float-child-expected.txt: Added.
* fast/block/margin-collapse/self-collapsing-block-with-overflow-hidden-and-float-child.html: Added.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@139337 268f45cc-cd09-0410-ab3c-d52691b4dbfc